PR feedback: Move cross-policy grant access methods to a separate module #627
