+++ app/policies/grant_submission/submission_policy.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
-class GrantSubmission::SubmissionPolicy < GrantPolicy
+class GrantSubmission::SubmissionPolicy < ApplicationPolicy
   attr_reader :user, :grant
+  include GrantRoleAccess
 
   def initialize(context, record)
     @user   = context.user
